Moving the headers to a common folder. Jazzy has a bug where umbrella headers for headers in multiple folders don't work correctly.
