+++ Source/ObjC/Blockly.h
@@ -22,8 +22,8 @@ FOUNDATION_EXPORT double BlocklyVersionNumber;
 FOUNDATION_EXPORT const unsigned char BlocklyVersionString[];
 
 // Public headers for Blockly
-#import <Blockly/BKYBlockJSONFile.h>
-#import <Blockly/BKYEdgeInsets.h>
-#import <Blockly/BKYLayoutConfigStructs.h>
-#import <Blockly/BKYWorkbenchViewController.h>
-#import <Blockly/BKYWorkspaceUnits.h>
+#import "BKYBlockJSONFile.h"
+#import "BKYEdgeInsets.h"
+#import "BKYLayoutConfigStructs.h"
+#import "BKYWorkbenchViewController.h"
+#import "BKYWorkspaceUnits.h"
